Clean up syntax errors
Fix syntax error reporting by using bison's built in error messages. Remove
the hand rolled recovery mechanism. This means it no longer supports bison <3.0
but it also means it works as is on both 3.0 and 3.6+.   Set errors to verbose.
Bison 3.6+ uses a type named context which was getting covered by the NscContext
name and causing compilation to fail on the generated code. Change that to "ctxt".